Move `maker_client.ts` to `taker/`
